refactor: extract TopLevelUnionFlattener and LongPressMetadataDetector
Split two pure, independently-testable seams out of oversized orchestrators
(issue #2659), with no behavior change:

- toolRegistry.ts: schema-flattening (~175 lines) -> src/server/TopLevelUnionFlattener.ts;
  flattenTopLevelUnion is re-exported for backward compatibility.
- TapOnElement.ts: long-press hierarchy analysis (~140 lines) ->
  src/features/action/LongPressMetadataDetector.ts, injected as a collaborator.

Adds direct unit tests for both modules.
